update action order

+++ Eshop-OP/routes/web.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\BrandsController;
 use App\Http\Controllers\AccountsController;
 use App\Http\Controllers\AccountController;
 use App\Http\Controllers\CartsController;
+use App\Http\Controllers\InvoicesController;
 
 
 
@@ -77,36 +78,7 @@ Route::get('/content', function () {
 Route::get('/product/create', function () {
     return view('product/create');
 });
-Route::prefix('/admin')->group(function (){
-    
-    Route::get('/create', [ProductsController::class, 'create'])->name('admin.create');
-    Route::get('/', [ProductsController::class, 'index'])->name('admin.index');
-    Route::get('/{id}', [ProductsController::class, 'show'])->name('admin.show');
-    Route::post('/', [ProductsController::class, 'store'])->name('admin.store');
-    Route::get('/edit/{id}', [ProductsController::class, 'edit'])->name('admin.edit');
-    Route::patch('/{id}', [ProductsController::class, 'update'])->name('admin.update');
-    Route::delete('/{id}', [ProductsController::class, 'destroy'])->name('admin.destroy');
-});
 
-Route::prefix('/category')->group(function (){
-    Route::get('/create', [CategorysController::class, 'create'])->name('category.create');
-    Route::get('/', [CategorysController::class, 'index'])->name('category.index');
-    Route::get('/{id}', [CategorysController::class, 'show'])->name('category.show');
-    Route::post('/', [CategorysController::class, 'store'])->name('category.store');
-    Route::get('/edit/{id}', [CategorysController::class, 'edit'])->name('category.edit');
-    Route::patch('/{id}', [CategorysController::class, 'update'])->name('category.update');
-    Route::delete('/{id}', [CategorysController::class, 'destroy'])->name('category.destroy');
-});
-
-Route::prefix('/brand')->group(function (){
-    Route::get('/create', [BrandsController::class, 'create'])->name('brand.create');
-    Route::get('/', [BrandsController::class, 'index'])->name('brand.index');
-    Route::get('/{id}', [BrandsController::class, 'show'])->name('brand.show');
-    Route::post('/', [BrandsController::class, 'store'])->name('brand.store');
-    Route::get('/edit/{id}', [BrandsController::class, 'edit'])->name('brand.edit');
-    Route::patch('/{id}', [BrandsController::class, 'update'])->name('brand.update');
-    Route::delete('/{id}', [BrandsController::class, 'destroy'])->name('brand.destroy');
-});
 
 
 Route::get('/login', [AccountController::class, 'index'])->name('index');
@@ -124,32 +96,41 @@ Route::get('/dashboard', function () {
 // Route::get('/create-product-type', [CategorysController::class, 'formCategory'])->name('create-product-type');
 // Route::post('/create-product-type', [CategorysController::class, 'create'])->name('create');
 
-Route::prefix('/product')->group(function (){
-    Route::get('/create', [ProductsController::class, 'create'])->name('product.create');
-    Route::get('/', [ProductsController::class, 'index'])->name('product.index');
-    Route::get('/{id}', [ProductsController::class, 'show'])->name('product.show');
-    Route::post('/', [ProductsController::class, 'store'])->name('product.store');
-    Route::get('/edit/{id}', [ProductsController::class, 'edit'])->name('product.edit');
-    Route::patch('/{id}', [ProductsController::class, 'update'])->name('product.update');
-    Route::delete('/{id}', [ProductsController::class, 'destroy'])->name('product.destroy');
+
+// Route JIDUY
+Route::prefix('/admin/product')->group(function (){
+    
+    Route::get('/create', [ProductsController::class, 'create'])->name('admin.product.create');
+    Route::get('/', [ProductsController::class, 'index'])->name('admin.product.index');
+    Route::get('/{id}', [ProductsController::class, 'show'])->name('admin.product.show');
+    Route::post('/', [ProductsController::class, 'store'])->name('admin.product.store');
+    Route::get('/edit/{id}', [ProductsController::class, 'edit'])->name('admin.product.edit');
+    Route::patch('/{id}', [ProductsController::class, 'update'])->name('admin.product.update');
+    Route::delete('/{id}', [ProductsController::class, 'destroy'])->name('admin.product.destroy');
 });
 
-Route::prefix('/category')->group(function (){
-    Route::get('/create', [CategorysController::class, 'create'])->name('category.create');
-    Route::get('/', [CategorysController::class, 'index'])->name('category.index');
-    Route::get('/{id}', [CategorysController::class, 'show'])->name('category.show');
-    Route::post('/', [CategorysController::class, 'store'])->name('category.store');
-    Route::get('/edit/{id}', [CategorysController::class, 'edit'])->name('category.edit');
-    Route::patch('/{id}', [CategorysController::class, 'update'])->name('category.update');
-    Route::delete('/{id}', [CategorysController::class, 'destroy'])->name('category.destroy');
+Route::prefix('/admin/category')->group(function (){
+    Route::get('/create', [CategorysController::class, 'create'])->name('admin.category.create');
+    Route::get('/', [CategorysController::class, 'index'])->name('admin.category.index');
+    Route::get('/{id}', [CategorysController::class, 'show'])->name('admin.category.show');
+    Route::post('/', [CategorysController::class, 'store'])->name('admin.category.store');
+    Route::get('/edit/{id}', [CategorysController::class, 'edit'])->name('admin.category.edit');
+    Route::patch('/{id}', [CategorysController::class, 'update'])->name('admin.category.update');
+    Route::delete('/{id}', [CategorysController::class, 'destroy'])->name('admin.category.destroy');
 });
 
-Route::prefix('/brand')->group(function (){
-    Route::get('/create', [BrandsController::class, 'create'])->name('brand.create');
-    Route::get('/', [BrandsController::class, 'index'])->name('brand.index');
-    Route::get('/{id}', [BrandsController::class, 'show'])->name('brand.show');
-    Route::post('/', [BrandsController::class, 'store'])->name('brand.store');
-    Route::get('/edit/{id}', [BrandsController::class, 'edit'])->name('brand.edit');
-    Route::patch('/{id}', [BrandsController::class, 'update'])->name('brand.update');
-    Route::delete('/{id}', [BrandsController::class, 'destroy'])->name('brand.destroy');
+Route::prefix('/admin/brand')->group(function (){
+    Route::get('/create', [BrandsController::class, 'create'])->name('admin.brand.create');
+    Route::get('/', [BrandsController::class, 'index'])->name('admin.brand.index');
+    Route::get('/{id}', [BrandsController::class, 'show'])->name('admin.brand.show');
+    Route::post('/', [BrandsController::class, 'store'])->name('admin.brand.store');
+    Route::get('/edit/{id}', [BrandsController::class, 'edit'])->name('admin.brand.edit');
+    Route::patch('/{id}', [BrandsController::class, 'update'])->name('admin.brand.update');
+    Route::delete('/{id}', [BrandsController::class, 'destroy'])->name('admin.brand.destroy');
+});
+
+Route::prefix('/admin/order')->group(function (){
+route::get('/', [InvoicesController::class, 'index'])->name('admin.order.index');
+route::get('/pending-approval', [InvoicesController::class, 'orderPendingApproval'])->name('admin.order.orderPendingApproval');
+route::get('/approval', [InvoicesController::class, 'orderApproval'])->name('admin.order.orderApproval');
 });
